plane works with shadows

+++ includes/errors.h
@@ -16,7 +16,7 @@ void check_sphere(t_figure *figure);
 void error(const char *string, int i);
 void check_color(t_color *color);
 void check_elements(t_data *data);
-int is_outrange_of_float(double number);
+int is_outrange_of_float(float number);
 void check_correct_file_extension(char *filename);
